Disentangle the pending_processor code
