+++ src/output.hpp
@@ -10,27 +10,12 @@
  * Associated tags: name, type etc.
 */
 
-#include <stack>
-
 #include <osmium/thread/pool.hpp>
 
 #include "options.hpp"
 
-struct expire_tiles;
-struct id_tracker;
 struct middle_query_t;
 class db_copy_thread_t;
-
-struct pending_job_t
-{
-    osmid_t osm_id;
-    size_t output_id;
-
-    pending_job_t() : osm_id(0), output_id(0) {}
-    pending_job_t(osmid_t id, size_t oid) : osm_id(id), output_id(oid) {}
-};
-
-using pending_queue_t = std::stack<pending_job_t>;
 
 class output_t
 {
